fixed scripts and also only cleaning db now if specified via args
lib.rs in main folder is required otherwise it doesnt compile

+++ server/files/server_data/db_server/createdb.sh
@@ -8,10 +8,13 @@ until docker exec postgres pg_isready -U admin -d postgres; do
     sleep 1
 done
 
-echo "Cleaning old db"
-docker exec postgres psql -U admin -d postgres -c "DROP DATABASE IF EXISTS rchats;"
-docker exec postgres psql -U admin -d postgres -c "CREATE DATABASE rchats;"
+if [[ "$1" = "clean" ]]; then
+    echo "Cleaning old db"
+    docker exec postgres psql -U admin -d postgres -c "DROP DATABASE IF EXISTS rchats;"
+    docker exec postgres psql -U admin -d postgres -c "CREATE DATABASE rchats;"
+    echo "Executing script to build db in postgres"
+    docker cp create_db.sql postgres:/create_db.sql
+    docker exec postgres psql -U admin -d rchats -f /create_db.sql
+fi
 
-echo "Executing script to build db in postgres"
-docker cp create_db.sql postgres:/create_db.sql
-docker exec postgres psql -U admin -d rchats -f /create_db.sql
+
